# Conflicts:
#	bundles/AdminBundle/Controller/Admin/DataObject/DataObjectController.php

+++ bundles/AdminBundle/Controller/Admin/Document/DocumentController.php
@@ -81,7 +81,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
             return $this->adminJson($data);
         }
 
-        return $this->adminJson(['success' => false, 'message' => 'missing_permission']);
+        throw $this->createAccessDeniedHttpException();
     }
 
     /**
@@ -242,7 +242,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
                     case 'page':
                         $document = Document\Page::create($request->get('parentId'), $createValues, false);
                         $document->setTitle($request->get('title', null));
-                        $document->setProperty('navigation_name', 'text', $request->get('name', null), false);
+                        $document->setProperty('navigation_name', 'text', $request->get('name', null), false, false);
                         $document->save();
                         $success = true;
                         break;
@@ -356,7 +356,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
 
             $deletedItems = [];
             foreach ($documents as $document) {
-                $deletedItems[] = $document->getRealFullPath();
+                $deletedItems[$document->getId()] = $document->getRealFullPath();
                 if ($document->isAllowed('delete') && !$document->isLocked()) {
                     $document->delete();
                 }
@@ -365,7 +365,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
             return $this->adminJson(['success' => true, 'deleted' => $deletedItems]);
         } elseif ($request->get('id')) {
             $document = Document::getById($request->get('id'));
-            if ($document->isAllowed('delete')) {
+            if ($document && $document->isAllowed('delete')) {
                 try {
                     if ($document->isLocked()) {
                         throw new \Exception('prevented deleting document, because it is locked: ID: ' . $document->getId());
@@ -381,7 +381,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
             }
         }
 
-        return $this->adminJson(['success' => false, 'message' => 'missing_permission']);
+        throw $this->createAccessDeniedHttpException();
     }
 
     /**
@@ -931,8 +931,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
                 }
             } else {
                 Logger::error('could not execute copy/paste because of missing permissions on target [ ' . $targetId . ' ]');
-
-                return $this->adminJson(['success' => false, 'message' => 'missing_permission']);
+                throw $this->createAccessDeniedHttpException();
             }
         }
 
@@ -1172,7 +1171,7 @@ class DocumentController extends ElementControllerBase implements EventedControl
             return $this->adminJson($nodeConfig);
         }
 
-        return $this->adminJson(['success' => false, 'message' => 'missing_permission']);
+        throw $this->createAccessDeniedHttpException();
     }
 
     /**
@@ -1449,6 +1448,14 @@ class DocumentController extends ElementControllerBase implements EventedControl
         $targetDocument = Document::getByPath($request->get('targetPath'));
 
         if ($sourceDocument && $targetDocument) {
+            if (empty($sourceDocument->getProperty('language'))) {
+                throw new \Exception(sprintf('Source Document(ID:%s) Language(Properties) missing', $sourceDocument->getId()));
+            }
+
+            if (empty($targetDocument->getProperty('language'))) {
+                throw new \Exception(sprintf('Target Document(ID:%s) Language(Properties) missing', $sourceDocument->getId()));
+            }
+
             $service = new Document\Service;
             if ($service->getTranslationSourceId($targetDocument) != $targetDocument->getId()) {
                 throw new \Exception('Target Document already linked to Source Document ID('.$service->getTranslationSourceId($targetDocument).'). Please unlink existing relation first.');
